test: move new delete-preserves-repo tests to server_cli2.py
The legacy lemonade-server shim no longer accepts pull/delete/import,
so tests using those subcommands must run against the new lemonade
binary in server_cli2.py.

Moves test_013/test_014 (added in this branch) to server_cli2.py as
test_091/test_092, alongside the helper functions and constants they
depend on.
